refactor(observability): downshift server metrics/prometheus surface to observability (P4a Batch 1a)
