+++ aragora/events/cross_subscribers/handlers/culture.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
 
 # Import metrics stubs - will be overwritten if metrics available
 try:
-    from aragora.server.prometheus_cross_pollination import record_km_outbound_event
+    from aragora.observability.prometheus_cross_pollination import record_km_outbound_event
 except ImportError:
 
     def record_km_outbound_event(target: str, event_type: str) -> None:
